fix issue with teachers not being able to authenticate
I do not know what the problem was for sure, but I removed the duplicate route for "/" from router.js and placed the code for creating a new teacher user in the redirectUser route. For some reason, the route for "/" was never even being reached, even when explicitly typed into the address bar, which is why authenticating teachers not already in the database did not work.